pkt_steer_module: Introduce idle core activation and add relevant stats
When no busy core can be found (either because no core is busy or all busy cores
ar overloaded) IAPS will actively activate a new idle core, when possible.
Before, the backup core would be chosen at random, so even though cores were
checked and rejected for being overloaded, not finding a suitable target could
steer a packet to an overloaded core anyways, which seems to cause performance
regressions at higher connection counts.

At module insertion, every core is put on the idle_backlog. It is to note that
cores are being inserted at the tail, so the initial idle_backlog will have the
structure 0->1->2->...->max(CPU).
When returning an idle core to idle_backlog, it is inserted at the head.
This is supposed to limit the cores that can be activated. IAPS still has a limit of
maximum busy cores, so idle activation will not activate a new core if the number is
already exceeded. That means if the number of maximum CPUs is 8, only cores 0~7 will
be activated.

For now, core 0 will not be put up as idle, which is a very static solution. This will
need to be fixed soon to include the provided min and max cpu values.

For now, it first needs to be established whether this feature is worth keeping.

This patch also introduces the stats noIdle, noIdleLB, and incorrectIdle, to keep track of
whether idle activation actually works and for what reasons it fails.

noIdle -> idle_backlog was empty
noIdleLB -> idle_backlog had entries but that max active cpu number didnt allow activation
incorrectIdle -> idle activation did not return an error, but the returned target is not
	   	 actually idle