pkt_steer_module: Abstract and split backup_core selection
With the idle activation now, backup_core selection has become more complex.
Some schemes may require idle core activation, some do not.
Also, the ones that need idle core activation still need to choose a backup_core
in case no idle core is found.

backup_core selection is split into two steps: proposal and decision.
The proposal is happens as before, where based on the selection requirement, a
core is selected that will be taken if no busy core will be found. However, as addition,
the selection algorithms that use idle activation mark that their backup choice
should be overwritten with an available idle core.
Then, when the algorithm reaches the point when it is clear that no busy core is available,
the final backup_core decision is made. If the algorithm does not rely on idle core activation,
choose the proposed backup_core. Otherwise, try to find an idle core first.