pkt_steer_module: Add option to disable util-based scaling
Now that it is possible to scale up active core number based on the available
cores' load, it should be able to handle appropriate scaling up and down without
considering CPU utilization. In IAPS+LB the limit on active cores could be a
limiting factor, so lifting it could be beneficial.