pkt_steer_module: Control CPUs entering busy and idle backlog
After introducing the separate idle and busy backlogs, there were
some issues where packets would be steered to cores that were not
configured for IAPS. The reason was that the IAPS core set is configurable,
but after configuration changes, the backlogs are not updated.

This was fixed by introducing core validity checks when inserting into
any of the backlogs. This keeps invalid CPUs out of the backlogs,
while allowing CPUs that were previously configured but became invalid
to still be removed from the backlogs.