pkt_steer_module: Add core selection latency infra
This commit introduces the necessary infrastructure to measure the latency
of the core selection during software-based packet steering.
Since core selection for RPS and RFS also runs through the module, this
can not only be used for IAPS.

Currently, the latency measurement granularity is static. A histogram with 20 bins
and a granularity of 50ns is created.

Latency measures are off by default and all relevant measurements are reset on every
activation, meaning that independent measurements are possible without having to
remove and reinsert the module.

Currently, the histogram data is printed when latency measurement is disabled.
Future commits will integrate the latency data into the proc file.