Experimental Scripts: Wrapper for main experiment script
The original script takes a lot of arguments that are not always
necessary, but since the arguments are taken positionally, they all
need to be listed every time.
To improve this, without running risk of messing up the original
script, this wrapper takes all the same arguments as arguments
(but with prefixing), so that it is easier to run the experiment.